domainpermissions can be sorted alphabetically or chronologically delete cache for morph-browser / webapps add settings page link to chrome of webapps
