+++ src/app/browserapplication.cpp
@@ -39,6 +39,7 @@
 #include "domain-settings-sorted-model.h"
 #include "domain-settings-user-agents-model.h"
 #include "favicon-fetcher.h"
+#include "file-operations.h"
 #include "input-method-handler.h"
 #include "meminfo.h"
 #include "mime-database.h"
@@ -97,10 +98,11 @@ QString BrowserApplication::inspectorHost() const
         return new type(); \
     }
 
-MAKE_SINGLETON_FACTORY(MemInfo)
-MAKE_SINGLETON_FACTORY(MimeDatabase)
 MAKE_SINGLETON_FACTORY(DomainPermissionsModel)
 MAKE_SINGLETON_FACTORY(DomainSettingsModel)
+MAKE_SINGLETON_FACTORY(FileOperations)
+MAKE_SINGLETON_FACTORY(MemInfo)
+MAKE_SINGLETON_FACTORY(MimeDatabase)
 MAKE_SINGLETON_FACTORY(UserAgentsModel)
 
 bool BrowserApplication::initialize(const QString& qmlFileSubPath
@@ -185,6 +187,7 @@ bool BrowserApplication::initialize(const QString& qmlFileSubPath
     qmlRegisterSingletonType<DomainSettingsModel>(uri, 0, 1, "DomainSettingsModel", DomainSettingsModel_singleton_factory);
     qmlRegisterType<DomainSettingsSortedModel>(uri, 0, 1, "DomainSettingsSortedModel");
     qmlRegisterType<FaviconFetcher>(uri, 0, 1, "FaviconFetcher");
+    qmlRegisterSingletonType<FileOperations>(uri, 0, 1, "FileOperations", FileOperations_singleton_factory);
     qmlRegisterSingletonType<MemInfo>(uri, 0, 1, "MemInfo", MemInfo_singleton_factory);
     qmlRegisterSingletonType<MimeDatabase>(uri, 0, 1, "MimeDatabase", MimeDatabase_singleton_factory);
     qmlRegisterType<SessionStorage>(uri, 0, 1, "SessionStorage");
